feat: collide players
