+++ rustyhack_client/src/consts.rs
@@ -1,6 +1,6 @@
-pub const VIEWPORT_WIDTH: u32 = 41;
-pub const VIEWPORT_HEIGHT: u32 = 15;
-pub const TARGET_FPS: u32 = 10;
-pub const LOG_NAME: &str = "rustyhack_client.log";
-pub const GAME_TITLE: &str = "Rustyhack MMO";
-pub const VALID_NAME_REGEX: &str = "^[[:alpha:]]+$";
+pub(crate) const VIEWPORT_WIDTH: u32 = 41;
+pub(crate) const VIEWPORT_HEIGHT: u32 = 15;
+pub(crate) const TARGET_FPS: u32 = 10;
+pub(crate) const LOG_NAME: &str = "rustyhack_client.log";
+pub(crate) const GAME_TITLE: &str = "Rustyhack MMO";
+pub(crate) const VALID_NAME_REGEX: &str = "^[[:alpha:]]+$";
