+++ rustyhack_lib/src/lib.rs
@@ -1,6 +1,3 @@
-pub use background_map::character_map;
-pub use background_map::tiles;
-
 pub mod background_map;
 pub mod consts;
 pub mod ecs;
