+++ rustyhack_server/src/consts.rs
@@ -1,5 +1,5 @@
 use std::time::Duration;
 
-pub const LOG_NAME: &str = "rustyhack_server.log";
-pub const ENTITY_UPDATE_TICK: Duration = Duration::from_millis(50);
-pub const LOOP_TICK: Duration = Duration::from_millis(10);
+pub(crate) const LOG_NAME: &str = "rustyhack_server.log";
+pub(crate) const ENTITY_UPDATE_TICK: Duration = Duration::from_millis(50);
+pub(crate) const LOOP_TICK: Duration = Duration::from_millis(10);
